# Refactor golang encode and decode