This refactors the `Encode()` and `Decode()` functions in order to improve readability.

The `Decode()` function gets a performance bonus - almost 10% faster, dropping from 285.3 ns/op to 262.5 ns/op.

Before:
```
BenchmarkEncode
BenchmarkEncode-14      31356085                34.69 ns/op           16 B/op          1 allocs/op
BenchmarkDecode
BenchmarkDecode-14       4183998               285.3 ns/op            24 B/op          1 allocs/op
```

After:
```
BenchmarkEncode
BenchmarkEncode-14      34198537                35.92 ns/op           16 B/op          1 allocs/op
BenchmarkDecode
BenchmarkDecode-14       4551160               262.5 ns/op            24 B/op          1 allocs/op
```